renamed asyncParser to warcIterator
