+++ lib/parsers/autoWARCParser.js
@@ -47,7 +47,7 @@ class AutoWARCParser extends EventEmitter {
     this._onError = this._onError.bind(this)
 
     if (typeof Symbol.asyncIterator !== 'undefined') {
-      const warcRecordIterator = require('./asyncParser')
+      const warcRecordIterator = require('./warcIterator')
       /**
        * @returns {AsyncIterator<WARCRecord>}
        */
