+++ lib/parsers/warcGzParser.js
@@ -42,7 +42,7 @@ class WARCGzParser extends EventEmitter {
     this._onError = this._onError.bind(this)
     this._onEnd = this._onEnd.bind(this)
     if (typeof Symbol.asyncIterator !== 'undefined') {
-      const warcRecordIterator = require('./asyncParser')
+      const warcRecordIterator = require('./warcIterator')
       /**
        * @returns {AsyncIterator<WARCRecord>}
        */
