+++ lib/parsers/warcIterator.js
