+++ lib/parsers/warcParser.js
@@ -41,7 +41,7 @@ class WARCParser extends EventEmitter {
     this._onError = this._onError.bind(this)
     this._onEnd = this._onEnd.bind(this)
     if (typeof Symbol.asyncIterator !== 'undefined') {
-      const warcRecordIterator = require('./asyncParser')
+      const warcRecordIterator = require('./warcIterator')
       /**
        * @returns {AsyncIterator<WARCRecord>}
        */
